Change move function to collapse all relevant squares when measuring
For this I added a hash, to keep track of which pieces might be entangled. Therefore, when a measure is needed for a square, we collapse all pieces that are entangled with that square and change the classical information accordingly. This is basically a fundamental change in the ruleset, but I felt it was needed to fix how split moves and entanglement work within the game.

Fix string parsing of moves not working for non 4x4 boards in the test game. Add NONE color type for NullPiece. Fix incorrect assert function being used in one test case.