Add quantum statistical tests and movement tests for the king
A quantum test engine was created to perform statistical testing of the different quantum moves. Because statistical testing has to be run multiple times to obtain meaningful results, it takes much longer than standard testing. As such, it might be useful to run only one type of test at a time. To make this easier there are now different directories for each type of test.