Add slide move and bishop, rook and queen pieces
Add all the needed functionality to Board, Piece and Point classes to be able to perform slide moves, as well as their respective tests. Split/merge slide not included in this commit.

Add an automated check to see if the game is over.
Add an automated construction of an initial board state. Since pawn movement is not yet done, pawns act as inmovable blockers.
Fix index-coordinate conversion not working for non squared boards.