Fix slide not properly collapsing the states sometimes
This resulted in incorrect board results when collapsing multiple times in a row. To solve it we previosly (32d680f35a34dd6b99453705787fce79fb436e9a) introduced the method is_piece_collapsed, but this method didn't properly collapse the quantum states.

The solution introduced in this commit is to add a reset operation (followed by X if needed) after collapsing to make sure that the qubit is properly collapsed, both classically and in the quantum circuit. Therefore the method is_piece_collapsed is deprecated.

We also make sure to collapse the path and the source at the same time. For this we introduced an optimization of collapse_path and entangle_path to only make one measure, and added an argument to also collapse the source or the target.