Add multiple game modes
Most of these game modes are not playable using QiskitEngine. This is because the max number of simulable qubits in a moder personal computer is around 25 and a M*N qchess board uses M*N+9. The more entangled the pieces become the harder it gets to simulate the circuit and measure it.

Fix incorrect indexing of pieces in QiskitEngine method does_slide_violate_double_occupancy.