Fix pieces being passed by reference instead of by value when moving
This could cause problems because if a piece is copied while moving (split, entangle, etc) then any change to one of the pieces would also take place in the other, which is not intended behavior. Since copying a pice is a very cheap operation, we just copy everytime a move happens just to be sure.