Restart the circuit when every piece is collapsed
This is the only reliable way I could think of (that doesn't involve state simulation) to improve performance. To accomplish it we add a piece.collapsed parameter that keeps track of the piece collapsed status. It's set to True only when measuring, and it's set to False everytime the piece is split.

This parameter is also used to raise an 'Invalid move' error when blocked by a collapsed piece.

Fix qflag not being reset when the piece is collapsed. To reset it, since we can't be 100% sure of the original qflag of a piece, we select it at random from the qflag used to collapse all the pieces.