Add pawn promotion
Fix pawns not checking if target piece or path had collapsed pieces in them.

The circuit now resets everytime collapse_by_flag is called (if all pieces are collapsed), even if no new piece was collapsed during that call. It's more efficient to reset the circuit than to keep a bunch of useless (since we already know the position of all the pieces) qubit operations.