Fix merge not working properly in non-trivial cases
Merge worked properly only when the piece was a superposition of two pieces that had been split only once, which is the easiest case (trivial). But when merge was done with pieces that are not split, or that come from different pieces, or that have different probability amplitudes, merge output incorrect results.

The fix is rather simple, but now merge will set certain pieces to non null even though those pieces have a 0% chance of existing. This is the tradeoff that was made, to keep the classical part of the program scalable.

It's also worth mentioning that split/merge behave differently depending on the order of targets/sources. This may seem incorrect (and that is up for debate), but it is a consecuence of how the circuits are designed. With this merge change it's more clear where the pieces might end up depending on that order.

Fix merge not checking properly if the piece's move was slide (was checking if the function existed instead of calling the function).

Refactor board to qchess in some tests.