Minor improvements to advanced_split and slide tutorials
Add the same scenario but with a black king in slide_tutorial to show the player that you can entangle pieces of different colors.

Move both kings one square to the left in advanced_split_jump_tutorial, effectively forcing the player to split the first king closer to the other so that the second movement can always be performed.